+++ src/lib/check_codeowners.rb
@@ -1,0 +1,7 @@
+require_relative './check_codeowners/check_runner'
+require_relative './check_codeowners/checks'
+require_relative './check_codeowners/get_options'
+require_relative './check_codeowners/individual_pattern_checker'
+require_relative './check_codeowners/multi_git_ls_runner'
+require_relative './check_codeowners/reports'
+require_relative './check_codeowners/repository'
