+++ src/lib/check_codeowners/repository/valid_owners.rb
@@ -1,0 +1,25 @@
+module CheckCodeowners
+  module Repository
+    class ValidOwners
+      def initialize(path)
+        @path = path
+      end
+
+      def valid_owners
+        @valid_owners ||= parse
+      end
+
+      private
+
+      attr_reader :path
+
+      def parse
+        begin
+          IO.readlines(path).map(&:chomp)
+        rescue Errno::ENOENT
+          nil
+        end
+      end
+    end
+  end
+end
